Fixed express implementation
