+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nomplate",
-  "version": "1.1.10",
+  "version": "1.1.11",
   "description": "Nomplate: Node template engine",
   "main": "index.js",
   "repository": {
